tests: make `runin()` accept a MIP version
Make `runIn()` and `RunInCondTrigger` accept a MIP version to use when
creating the graph, instead of using 0.  Update existing users to pass a
hardcoded 0.

Add a `forEachMipVersion()` utility function, which calls a callback
once for each possible MIP version.  This is intended to help code that
wants to try testing variations of the same thing for all possible MIP
version.

Change-Id: Ic1d42d20b4eef872106f216fea9d2e883cb33775
Signed-off-by: Simon Marchi <simon.marchi@efficios.com>