debug-info: Compile as C++
Justifications
==============

Leave `crc32.h` and `crc32.c` as C code: The file is code
external to the project and it is typical to never modify these files.
However, when compiling alonside C++ code we must explicitly specify
that it is C code so that the compiler knows to refer to function
names with the non-mangled symbol name. Specifying that `crc32.h` is C
in the file was less intrusive compared to translating the whole file
to C++.

Change-Id: I5e8823ceca994c035de6ebba5df22b2c3d11a02b
Signed-off-by: Erica Bugden <ebugden@efficios.com>
Reviewed-on: https://review.lttng.org/c/babeltrace/+/13225
Reviewed-by: Simon Marchi <simon.marchi@efficios.com>