debug-info: Compile as C++
Justifications
==============

Add specification to compile `crc32.h` as C code: The file is code
external to the project and it is typical to never modify files.
However, a change needed to be made in order to compile the file
alongside the new C++ code. Specifying that `crc32.h` should be
compiled as C was less intrusive compared to translating the whole file
to C++.

Change-Id: I5e8823ceca994c035de6ebba5df22b2c3d11a02b
Signed-off-by: Erica Bugden <ebugden@efficios.com>