fd-cache: Include `glib.h`
The debug info filter files have been transitioned to C++ in large part
to allow automatic code formatting. However, simply auto-formatting the
code causes the build to fail. We are not rigorous with
"include what you use" so it's likely that the formatting adjusted the
order of includes and this resulted in a needed include being missing.

Include `glib.h` before auto-formatting the debug info C++ files to
avoid this build failure.

Change-Id: I64665841559a70d7bf28ddb27a58b7c3ba2bfb43
Signed-off-by: Erica Bugden <ebugden@efficios.com>