ccv: output reference clock class on error
On validation error, always output the clock class used as a reference.
In combination with the error type, the callers will be able to inspect
the reference clock class to obtain the details needed to generate the
appropriate error message.

The goal of this change is to avoid having to add many output parameters
when adding support for MIP 1, keeping things simpler.

Change-Id: I032cc706d586134a35aabd14d9fb00611b430ad0
Signed-off-by: Simon Marchi <simon.marchi@efficios.com>