lib: add bt_field_class_get_graph_mip_version()
Add bt_field_class_get_graph_mip_version, which returns the MIP version
for the graph the field class is in, for the same reason as explained in
fbc90342dade ("lib: add bt_trace_class_get_graph_mip_version") for trace
class.

Note to Phil: I don't know where to put the function declaration in the
header, perhaps a new section is needed.

Signed-off-by: Simon Marchi <simon.marchi@efficios.com>
Change-Id: I97b4a440020cb006bee9a79a5dce9cbade565914
Reviewed-on: https://review.lttng.org/c/babeltrace/+/13395
Reviewed-by: Philippe Proulx <eeppeliteloop@gmail.com>