bt2/setup.py.in: assume environment variables always exist
`setup.py` is always called from our Makefile, where we control which
environment variables are set.  Remove the code paths that handle the
cases where environment variables CC, LDFLAGS, CFLAGS and CPPFLAGS may
not exist.  This simplifies the number of cases to worry about, making
the code easier to reason about.

This requires passing the build flags to setup.py when using the install
command, but I don't think that's a problem.

Change-Id: I380d113824bddf6a423e34a681d904a844175100
Signed-off-by: Simon Marchi <simon.marchi@efficios.com>
Reviewed-on: https://review.lttng.org/c/babeltrace/+/13491
Reviewed-by: Michael Jeanson <mjeanson@efficios.com>
Tested-by: jenkins <jenkins@lttng.org>
Reviewed-by: Philippe Proulx <eeppeliteloop@gmail.com>