lib: record field class parents, log trace IR path of field classes
The initial motivation for this change came from the following
pre-condition failure message:

    Field is not set: addr=0x507000003820, is-set=0, is-frozen=0, class-addr=0x50b000016bf0, class-type=UNSIGNED_INTEGER

I think this could be made more user-friendly if the message gave the
path to the field.  After this change, the message also includes this:

    class-trace-ir-path=/<stream class 0>/<event class "my-event">/<event-payload>/variant-with-unsigned-integer-selector-field-selector

I'm not particularly attached to the specific syntax, but here's how it
works:

 ‣ levels are separated by forward slashes, like a Unix filesystem, a
   leading forward slash indicating an "absolute" path
 ‣ for struct members or variant options with a name, the name is
   printed "naked"
 ‣ there is no level printed for field classes used as array element or
   optional content
 ‣ for variant options without a name, the index is printed like
   this:

     <option 3>

 ‣ stream classes and event classes are printed in angle brackets, with
   their name, or their id, if the don't have a name, e.g.:

     <stream-class 3>
     <event-class "foo">

 ‣ scopes are printed in angle brackets too:

     <packet-context>
     <event-common-context>
     <event-specific-context>
     <event-payload>

 ‣ if the field class was not yet added to a scope, the path does not
   start with a forward slash, and begins with (taking suggestions for a
   better terminology here):

     <not connected>

To implement this, record in `bt_field_class` the parent field class,
and the index in the parent, if applicable.  For structure field classes
specifically, also record if it is being used as the root of a stream
class or event class scope.

Not that this only prints the path of the field classes in the trace IR,
not the path of fields.  I think the message could be even more precise
if we printed the path of fields, since it could include the specific
array index where an error occured.  For instance, if something bad
happened while working on a field that is the 3rd element of an array,
the path could be something like:

    trace-ir-path=/<stream 0>/<event "my-event">/<payload>/my-array-field/<2>

I haven't tried to implement this yet, but I think it can be future
work.  I think the code doing this for fields would be a bit redundant
with the field class feature, but I don't really see a way around it.

Change-Id: Ia44d9e0dff815bc804d6b15c03128f228d37b246
Signed-off-by: Simon Marchi <simon.marchi@efficios.com>