lib: remove `bt_field_class::part_of_trace_class`, improve some pre-condition checks
Now that field classes record their parent, implement the following
pre-condition checks:

 ‣ when creating an option field class:

     optional_field_class is not already part of another field class, an
     event class, or a stream class.

 ‣ when adding an option to a variant field class:

     option_field_class is not already part of another field class, an
     event class, or a stream class.

 ‣ when adding a member to a structure field class:

     member_field_class is not already part of another field class, an
     event class, or a stream class.

 ‣ when creating an array field class:

     element_field_class is not already part of another field class, an
     event class, or a stream class.

Remove the `bt_field_class::part_of_trace_class` flag which, in my
opinion, becomes unnecessary.  It is currently being used to check the
following assertion, when a scope of a stream or event class (as well as
in logging):

     field_class, or any of its contained field classes, is not already
     part of a stream class or of an event class.

However, in addition to the other pre-conditions above, if we simply
check that the structure field class used when setting the stream or
event class scope isn't used for anything already ("not already part of
another field class, an event class, or a stream class") - i.e. not
recursively - I think we are good.

Change-Id: I0ad5d12e808ecfe94fa4a63e4e23fbe0991a201c
Signed-off-by: Simon Marchi <simon.marchi@efficios.com>