bt2: pass expected type parameter to `_TraceClass._check_and_wrap_field_class()`
Add a generic type parameter to `_check_and_wrap_field_class()`, so
that it can assert that the type of the created field class is as
expected.  This also allows defining its return type to the right thing,
depending on the caller, removing a bunch of type errors in the callers.

Change-Id: I4d0d1ddf1135b8a2ede5a96f19b7b0938756e0a8
Signed-off-by: Simon Marchi <simon.marchi@efficios.com>