tests: add "complete" source
Move the plugin `bt_plugin_test_debug_info.py`, currently specific to
the `flt.lttng-utils.debug-info` tests, to `tests/utils/python`, to make
it available to other tests.  This plugin defines one source that sends
a big of everything.  This source has nothing specific to
`flt.lttng-utils.debug-info`, and I think it could be useful for other
tests.

Rename the file to `bt_plugin_complete.py` and the plugin to `complete`.

Do minor adjustments to the code to make it clean, typing-wise.

Change-Id: Ieb91ce5e7a8e70676f2fd4b7937c4347d5970620
Signed-off-by: Simon Marchi <simon.marchi@efficios.com>