bt2: move typing import to separate file
Our incredibly fragile little trick to import the typing module seems to
have a problem: it doesn't work to use the typing module in `utils.py`
itself.  At least, it seems like pyright doesn't recognize it as "the"
typing module, and so its logic for it doesn't kick in.

Move the import to a new file, so that `utils.py` can use it like other
files do.

Change-Id: I3ec9ad164c49242ba73b970b80b1b6f1e9266064
Signed-off-by: Simon Marchi <simon.marchi@efficios.com>