bt2: make `_check_type()` return the object
This will help with typing in some situation.  For instance, if a
parameter `x` is of type `typing.Optional[Foo]`, but in a given branch,
`x` must not be `None`, then you can use:

    x = bt2_utils._check_type(x, Foo)

to make the type checker understand that from that point, `x` is `Foo`,
not `typing.Optional[Foo]`.  This acts as a somewhat nicer and shorter
`assert x is not None`.

Change-Id: Id5f4a0ab8f87302d3f1d2e9ea0d385310fd6678b
Signed-off-by: Simon Marchi <simon.marchi@efficios.com>